+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add possibility to define width and height options for BalloonFeatureInfo
 - Add support for template strings in window state headerTitle. See [FeatureInfo](documentation/FEATURE_INFO.md#window-title)
 - Fix bug, where balloons were clipped to target
+- Fixes a bug where CollectionManager `removeActions` was not working as expected
 
 # 5.1.9